feat: implement episodesSlice
