+++ src/components/episodes/episodes.module.css
@@ -1,0 +1,53 @@
+.episodeDetail {
+  flex: 0 1 50%;
+}
+
+.episodesNavigation {
+  block-size: 100%;
+  min-inline-size: 10em;
+  margin-block-end: 1em;
+  overflow-y: auto;
+  overflow-x: hidden;
+
+  border-radius: var(--border-radius-input);
+  background-color: rgb(var(--input-color-background));
+
+  &::-webkit-scrollbar {
+    width: 7px;
+    height: 5px;
+  }
+  &::-webkit-scrollbar-button {
+    width: 0px;
+    height: 0px;
+  }
+  &::-webkit-scrollbar-thumb {
+    background: var(--color-scroll);
+    border-radius: 0px;
+  }
+  &::-webkit-scrollbar-thumb:hover {
+    background: color-mix(in srgb, var(--color-scroll) 90%, white 10%);
+  }
+  &::-webkit-scrollbar-thumb:active {
+    background: color-mix(in srgb, var(--color-scroll) 90%, white 10%);
+  }
+  &::-webkit-scrollbar-track {
+    background: var(--color-scroll-background);
+    border-radius: 5px;
+  }
+  &::-webkit-scrollbar-track:hover {
+    background: var(--color-scroll-background);
+  }
+  &::-webkit-scrollbar-track:active {
+    background: var(--color-scroll-background);
+  }
+  &::-webkit-scrollbar-corner {
+    background: transparent;
+  }
+}
+
+.navElements {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+}
