+++ src/components/main/main.module.css
@@ -6,78 +6,3 @@
   color: black;
   padding: 1em;
 }
-
-.episodeDetail {
-  flex: 0 1 50%;
-}
-
-.episodesNavigation {
-  block-size: 100%;
-  min-inline-size: 10em;
-  margin-block-end: 1em;
-  overflow-y: auto;
-  overflow-x: hidden;
-
-  border-radius: var(--border-radius-input);
-  background-color: rgb(var(--input-color-background));
-
-  &::-webkit-scrollbar {
-    width: 7px;
-    height: 5px;
-  }
-  &::-webkit-scrollbar-button {
-    width: 0px;
-    height: 0px;
-  }
-  &::-webkit-scrollbar-thumb {
-    background: var(--color-scroll);
-    border-radius: 0px;
-  }
-  &::-webkit-scrollbar-thumb:hover {
-    background: color-mix(in srgb, var(--color-scroll) 90%, white 10%);
-  }
-  &::-webkit-scrollbar-thumb:active {
-    background: color-mix(in srgb, var(--color-scroll) 90%, white 10%);
-  }
-  &::-webkit-scrollbar-track {
-    background: var(--color-scroll-background);
-    border-radius: 5px;
-  }
-  &::-webkit-scrollbar-track:hover {
-    background: var(--color-scroll-background);
-  }
-  &::-webkit-scrollbar-track:active {
-    background: var(--color-scroll-background);
-  }
-  &::-webkit-scrollbar-corner {
-    background: transparent;
-  }
-}
-
-.navElements {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: center;
-}
-
-@media (width < 1025px) {
-  .main {
-    --template-column: 1fr 1fr 1fr;
-    --row-height: 11em;
-  }
-}
-
-@media (width < 769px) {
-  .main {
-    --template-column: 1fr 1fr;
-    --row-height: 12em;
-  }
-}
-
-@media (width < 501px) {
-  .main {
-    --template-column: 1fr;
-    --row-height: 13em;
-  }
-}
